czwarty commit

+++ KolokwiumPoprawkowe2-GorskiBartosz/Struktury/zwierze.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include "zwierze.h"
-Zwierze initZwierze(const char *gatunek,const int masa,const unsigned int wiek){
+Zwierze initZwierze(const char *gatunek,const double masa,const unsigned int wiek){
     Zwierze temp;
     strcpy(temp.gatunek,gatunek);
     temp.masa = masa;
@@ -11,10 +11,10 @@ Zwierze initZwierze(const char *gatunek,const int masa,const unsigned int wiek){
 }
 void showZwierze(const Zwierze *z){
     printf("\nGatunek: %s",z->gatunek);
-    printf("\nMasa: %d",z->masa);
+    printf("\nMasa: %f",z->masa);
     printf("\nWiek: %d",z->wiek);
 }
-Zwierze tycieZwierzecia(const Zwierze *z,int dodajkg){
+Zwierze tycieZwierzecia(const Zwierze *z,double dodajkg){
     Zwierze temp;
     strcpy(temp.gatunek,z->gatunek);
     temp.masa = z->masa + dodajkg;
