+++ KolokwiumPoprawkowe2-GorskiBartosz/Struktury/zwierze.h
@@ -3,10 +3,10 @@
 #define MAX_LEN 20
 typedef struct Zwierze{
     char gatunek[MAX_LEN - 1];
-    int masa;
+    double masa;
     unsigned int wiek;
 }Zwierze;
-Zwierze initZwierze(const char *gatunek,const int masa,const unsigned int wiek);
-Zwierze tycieZwierzecia(const Zwierze *z,int dodajkg);
+Zwierze initZwierze(const char *gatunek,const double masa,const unsigned int wiek);
+Zwierze tycieZwierzecia(const Zwierze *z,double dodajkg);
 void showZwierze(const Zwierze *z);
 #endif
